增加接收端的丢包统计，修改feedback协议，将丢包统计加入到feedback协议当中

+++ estimator/rate_stat.c
@@ -49,7 +49,7 @@ static void rate_stat_erase(rate_stat_t* rate, int64_t now_ts)
 	if (rate->oldest_ts == -1)
 		return;
 	
-	new_oldest_ts = now_ts - rate->wnd_size + 1;
+	new_oldest_ts = (int)(now_ts - rate->wnd_size + 1);
 	if (new_oldest_ts <= rate->oldest_ts)
 		return;
 
@@ -83,7 +83,7 @@ void rate_stat_update(rate_stat_t* rate, size_t count, int64_t now_ts)
 		rate->oldest_ts = now_ts;
 	}
 
-	ts_offset = now_ts - rate->oldest_ts;
+	ts_offset = (int)(now_ts - rate->oldest_ts);
 	index = (rate->oldest_index + ts_offset) % rate->wnd_size;
 	
 	rate->buckets[index].sum += count;
@@ -102,11 +102,11 @@ int rate_stat_rate(rate_stat_t* rate, int64_t now_ts)
 
 	rate_stat_erase(rate, now_ts);
 
-	active_wnd_size = now_ts - rate->oldest_ts + 1;
+	active_wnd_size = (int)(now_ts - rate->oldest_ts + 1);
 	if (rate->sample_num == 0 || active_wnd_size <= 1 || (rate->sample_num == 1 && active_wnd_size <= rate->wnd_size))
 		return -1;
 
-	ret = rate->accumulated_count * rate->scale / active_wnd_size + 0.5;
+	ret = (int)(rate->accumulated_count * rate->scale / active_wnd_size + 0.5);
 
 	return ret;
 }
